Add configuration for database, lombok and example entity
Adding configuration for database. This method is native for spring
and will work out of the box with any type of database which includes
in memory, in file, or mysql server. Currently in the configuration
all the approaches are specified and the in memory database is
uncommented as the default approach.

Lombok plugin is applied to simplify redundant code. Now setters,
getters, constructors etc. can be created with simple annotation.

Also downloaded the javascript which was loaded from the web as
a local file to not require bandwith for the application to work.

How to create entities in the database is specified as commented
out code in LoginController#login

Signed-off-by: Martin Dekov <mvdekov@gmail.com>